Javadoc kommentek hozzáadva

+++ src/game_elements/entities/PowerUp.java
@@ -10,6 +10,9 @@ import java.io.IOException;
 import java.util.List;
 import java.util.ArrayList;
 
+/**
+ * A powerup osztály.
+ */
 public class PowerUp extends GameEntity {
 
     private static List<BufferedImage> powerupImageList;
@@ -42,6 +45,9 @@ public class PowerUp extends GameEntity {
         powerupInit();
     }
 
+    /**
+     * Az értékek inícializásáért felelős függvény.
+     */
     public void powerupInit() {
         speed = 5;
         powerupType = (int) (4*Math.random());
@@ -54,24 +60,39 @@ public class PowerUp extends GameEntity {
 
     }
 
+    /**
+     * Frissíti a powerup pozícióját és ellenőrzi, hogy még a határain belül tartózkodik-e.
+     */
     public void update() {
         yCoordinate += speed;
         if(yCoordinate > maxCoordinateOfY)
             isOutOfBounds = true;
     }
 
-
+    /**
+     * A megjelenítésért felelős függvény.
+     * @param g A megjelenítést végző Graphics2D objektum.
+     */
     public void draw(Graphics2D g) {
-        g.drawImage(puImage, xCoordinate, yCoordinate, null);
+        g.drawImage(puImage, (int) xCoordinate, (int) yCoordinate, null);
     }
 
+    /**
+     * A powerup határait visszaadó függvény.
+     * A powerup köré egy téglalapot von, ami a határait jelzi.
+     *
+     * @return Egy Rectangle objektumként a határát.
+     */
     public Rectangle getPowerUpBounds() {
-        return new Rectangle(xCoordinate, yCoordinate, size_x, size_y);
+        return new Rectangle((int) xCoordinate, (int) yCoordinate, size_x, size_y);
     }
 
+    /**
+     * A powerup típusát adja vissza int értékként.
+     *
+     * @return 0 - extra lézer, 1 - extra élet, 2 - pajzs, 3 - rövidebb újratöltődési idő.
+     */
     public int getType() {
         return powerupType;
     }
-
-
 }
